PRIu8 format specifier does not work in zephyr with some specific LIBC

+++ components/nn/src/nn/ip_address.c
@@ -20,7 +20,7 @@ const char* nn_ip_address_to_string(const struct nn_ip_address* address)
     memset(outputBuffer, 0, 40);
     if (address->type == NN_IPV4) {
         const uint8_t* ip = address->ip.v4;
-        sprintf(outputBuffer, "%" PRIu8 ".%" PRIu8 ".%" PRIu8 ".%" PRIu8, ip[0], ip[1], ip[2], ip[3]);
+        sprintf(outputBuffer, "%u.%u.%u.%u", (unsigned int)(ip[0]), (unsigned int)(ip[1]), (unsigned int)(ip[2]), (unsigned int)(ip[3]));
     } else if (address->type == NN_IPV6) {
         const uint8_t* ip = address->ip.v6;
         sprintf(outputBuffer, "%02x%02x:" "%02x%02x:" "%02x%02x:" "%02x%02x:" "%02x%02x:" "%02x%02x:" "%02x%02x:" "%02x%02x", ip[0], ip[1], ip[2], ip[3], ip[4], ip[5], ip[6], ip[7], ip[8], ip[9], ip[10], ip[11], ip[12], ip[13], ip[14], ip[15]);
